[AV2-941] upgraded to v3
