--- android/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="my_string">Just a simple string</string>
-</resources>
