--- android/src/main/res/values/styles.xml
@@ -1,3 +1,0 @@
-<resources>
-
-</resources>
