+++ rollup.config.js
@@ -2,15 +2,24 @@ import nodeResolve from '@rollup/plugin-node-resolve';
 
 export default {
   input: 'dist/esm/index.js',
-  output: {
-    file: 'dist/plugin.js',
-    format: 'iife',
-    name: 'capacitorPlugin', // TODO: change this
-    globals: {
-      '@capacitor/core': 'capacitorExports',
+  output: [
+    {
+      file: 'dist/plugin.js',
+      format: 'iife',
+      name: 'capacitorPlugin', // TODO: change this
+      globals: {
+        '@capacitor/core': 'capacitorExports',
+      },
+      sourcemap: true,
+      inlineDynamicImports: true,
     },
-    sourcemap: true,
-  },
+    {
+      file: 'dist/plugin.cjs.js',
+      format: 'cjs',
+      sourcemap: true,
+      inlineDynamicImports: true,
+    }
+  ],
   plugins: [
     nodeResolve({
       // allowlist of dependencies to bundle in
@@ -18,6 +27,5 @@ export default {
       resolveOnly: ['lodash'],
     }),
   ],
-  // context: ,
-  // moduleContext: ,
+  external: ['@capacitor/core'],
 };
